Refactor BH codebase to be compatible with C++20
