+++ BH/BH.vcxproj
@@ -174,6 +174,12 @@ if defined D2Path (
     <ClCompile Include="Modules\Gamefilter\Gamefilter.cpp" />
     <ClCompile Include="Modules\Item\Item.cpp" />
     <ClCompile Include="Modules\Item\ItemDisplay.cpp" />
+    <ClCompile Include="Modules\Item\ItemDisplay\Action.cpp" />
+    <ClCompile Include="Modules\Item\ItemDisplay\Condition.cpp" />
+    <ClCompile Include="Modules\Item\ItemDisplay\GemRune.cpp" />
+    <ClCompile Include="Modules\Item\ItemDisplay\ItemHelper.cpp" />
+    <ClCompile Include="Modules\Item\ItemDisplay\LookupCache.cpp" />
+    <ClCompile Include="Modules\Item\ItemDisplay\Rule.cpp" />
     <ClCompile Include="Modules\ItemMover\ItemMover.cpp" />
     <ClCompile Include="Modules\Maphack\Maphack.cpp" />
     <ClCompile Include="Modules\Module.cpp" />
@@ -235,6 +241,17 @@ if defined D2Path (
     <ClInclude Include="Modules\Gamefilter\Gamefilter.h" />
     <ClInclude Include="Modules\Item\Item.h" />
     <ClInclude Include="Modules\Item\ItemDisplay.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\Action.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\ColorConstants.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\Condition.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\ConditionType.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\GemRune.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\ItemHelper.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\ItemInfo.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\ItemProperty.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\LookupCache.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\Rule.h" />
+    <ClInclude Include="Modules\Item\ItemDisplay\UnitItemInfo.h" />
     <ClInclude Include="Modules\ItemMover\ItemMover.h" />
     <ClInclude Include="Modules\Maphack\Maphack.h" />
     <ClInclude Include="Modules\Module.h" />
