+++ BH/D2Stubs.h
@@ -10,7 +10,7 @@ DWORD __fastcall D2CLIENT_InitAutomapLayer(DWORD nLayerNo);
 DWORD __fastcall TestPvpFlag_STUB(DWORD planum1, DWORD planum2, DWORD flagmask);
 DWORD __fastcall D2CLIENT_GetLevelName_STUB(DWORD levelId);
 DWORD __fastcall D2CLIENT_GetUIVar_STUB(DWORD varno);
-CellFile* __fastcall D2CLIENT_LoadUiImage(CHAR* szPath);
+CellFile* __fastcall D2CLIENT_LoadUiImage(const CHAR* szPath);
 DWORD __fastcall D2CLIENT_ClickParty_ASM(RosterUnit* RosterUnit, DWORD Mode);
 void __fastcall D2CLIENT_PlaySound(int SoundNo);
 void* __stdcall D2WIN_LoadMpq(DWORD OrderFlag, const char* szDllFile, const char* szMpqFile, const char* szName, DWORD _1, BOOL(__stdcall* fnMPQMissing)());
