+++ BH/Modules/Item/ItemDisplay/ColorConstants.h
@@ -1,0 +1,8 @@
+
+#ifndef BH_MODULES_ITEM_ITEM_DISPLAY_COLOR_CONSTANTS_H_
+#define BH_MODULES_ITEM_ITEM_DISPLAY_COLOR_CONSTANTS_H_
+
+#define DEAD_COLOR        0xdead
+#define UNDEFINED_COLOR   0xbeef
+
+#endif  // BH_MODULES_ITEM_ITEM_DISPLAY_COLOR_CONSTANTS_H_
